BUG 4263: No test coverage for SfcProviderSfEntryDataListener

+++ sfc-provider/src/main/java/org/opendaylight/sfc/provider/SfcProviderSfEntryDataListener.java
@@ -63,55 +63,59 @@ public class SfcProviderSfEntryDataListener implements DataChangeListener {
 
                 // SF ORIGINAL
                 Map<InstanceIdentifier<?>, DataObject> dataOriginalDataObject = change.getOriginalData();
-
-                for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataOriginalDataObject.entrySet()) {
-                    if (entry.getValue() instanceof ServiceFunction) {
-                        ServiceFunction originalServiceFunction = (ServiceFunction) entry.getValue();
-                        LOG.debug("\n########## getOriginalConfigurationData {}  {}",
-                                originalServiceFunction.getType(), originalServiceFunction.getName());
+                if (dataOriginalDataObject != null) {
+                    for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataOriginalDataObject.entrySet()) {
+                        if (entry.getValue() instanceof ServiceFunction) {
+                            ServiceFunction originalServiceFunction = (ServiceFunction) entry.getValue();
+                            LOG.debug("\n########## getOriginalConfigurationData {}  {}",
+                                    originalServiceFunction.getType(), originalServiceFunction.getName());
+                        }
                     }
                 }
 
                 // SF CREATION
                 Map<InstanceIdentifier<?>, DataObject> dataCreatedObject = change.getCreatedData();
+                if (dataCreatedObject != null) {
+                    for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataCreatedObject.entrySet()) {
+                        if (entry.getValue() instanceof ServiceFunction) {
+                            ServiceFunction createdServiceFunction = (ServiceFunction) entry.getValue();
 
-                for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataCreatedObject.entrySet()) {
-                    if (entry.getValue() instanceof ServiceFunction) {
-                        ServiceFunction createdServiceFunction = (ServiceFunction) entry.getValue();
-
-                        if (!SfcProviderServiceTypeAPI.createServiceFunctionTypeEntryExecutor(createdServiceFunction)) {
-                            LOG.error("Failed to create service function type: {}", createdServiceFunction.getType());
+                            if (!SfcProviderServiceTypeAPI.createServiceFunctionTypeEntryExecutor(createdServiceFunction)) {
+                                LOG.error("Failed to create service function type: {}", createdServiceFunction.getType());
+                            }
                         }
                     }
                 }
 
                 // SF DELETION
                 Set<InstanceIdentifier<?>> dataRemovedConfigurationIID = change.getRemovedPaths();
-                for (InstanceIdentifier instanceIdentifier : dataRemovedConfigurationIID) {
-                    DataObject dataObject = dataOriginalDataObject.get(instanceIdentifier);
-                    if (dataObject instanceof ServiceFunction) {
-                        ServiceFunction originalServiceFunction = (ServiceFunction) dataObject;
+                if (dataRemovedConfigurationIID != null) {
+                    for (InstanceIdentifier instanceIdentifier : dataRemovedConfigurationIID) {
+                        DataObject dataObject = dataOriginalDataObject.get(instanceIdentifier);
+                        if (dataObject instanceof ServiceFunction) {
+                            ServiceFunction originalServiceFunction = (ServiceFunction) dataObject;
 
-                        if (!SfcProviderServiceTypeAPI.deleteServiceFunctionTypeEntryExecutor(originalServiceFunction)) {
-                            LOG.error("Failed to delete Service Function Type for SF: {}", originalServiceFunction.getName());
-                        }
-
-                /* Before removing RSPs used by this Service Function, we need to remove all
-                 * references in the SFF/SF operational trees
-                 */
-                        String sfName = originalServiceFunction.getName();
-                        List<String> rspList = SfcProviderServiceFunctionAPI.
-                                readServiceFunctionStateAsStringListExecutor(sfName);
-                        if ((rspList != null) && (!rspList.isEmpty())) {
-                            if (SfcProviderServiceFunctionAPI.deleteServiceFunctionStateExecutor(sfName)) {
-                            } else {
-                                LOG.error("{}: Failed to delete SF {} operational state",
-                                        Thread.currentThread().getStackTrace()[1], sfName);
+                            if (!SfcProviderServiceTypeAPI.deleteServiceFunctionTypeEntryExecutor(originalServiceFunction)) {
+                                LOG.error("Failed to delete Service Function Type for SF: {}", originalServiceFunction.getName());
                             }
-                            SfcProviderServiceForwarderAPI
-                                    .deletePathFromServiceForwarderStateExecutor(rspList);
 
-                            SfcProviderRenderedPathAPI.deleteRenderedServicePathsExecutor(rspList);
+                            /* Before removing RSPs used by this Service Function, we need to remove all
+                             * references in the SFF/SF operational trees
+                             */
+                            String sfName = originalServiceFunction.getName();
+                            List<String> rspList = SfcProviderServiceFunctionAPI.
+                                    readServiceFunctionStateAsStringListExecutor(sfName);
+                            if ((rspList != null) && (!rspList.isEmpty())) {
+                                if (SfcProviderServiceFunctionAPI.deleteServiceFunctionStateExecutor(sfName)) {
+                                } else {
+                                    LOG.error("{}: Failed to delete SF {} operational state",
+                                            Thread.currentThread().getStackTrace()[1], sfName);
+                                }
+                                SfcProviderServiceForwarderAPI
+                                        .deletePathFromServiceForwarderStateExecutor(rspList);
+
+                                SfcProviderRenderedPathAPI.deleteRenderedServicePathsExecutor(rspList);
+                            }
                         }
                     }
                 }
@@ -120,61 +124,61 @@ public class SfcProviderSfEntryDataListener implements DataChangeListener {
                 // SF UPDATE
                 Map<InstanceIdentifier<?>, DataObject> dataUpdatedConfigurationObject
                         = change.getUpdatedData();
-                for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataUpdatedConfigurationObject.entrySet()) {
-                    if ((entry.getValue() instanceof ServiceFunction) && (!(dataCreatedObject.containsKey(entry.getKey())))) {
+                if (dataUpdatedConfigurationObject != null && dataCreatedObject != null) {
+                    for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataUpdatedConfigurationObject.entrySet()) {
+                        if ((entry.getValue() instanceof ServiceFunction) && (!(dataCreatedObject.containsKey(entry.getKey())))) {
+                            DataObject dataObject = dataOriginalDataObject.get(entry.getKey());
+                            ServiceFunction originalServiceFunction = (ServiceFunction) dataObject;
+                            Object[] serviceFunctionObj = {originalServiceFunction};
+                            Class[] serviceFunctionClass = {ServiceFunction.class};
 
 
-                        DataObject dataObject = dataOriginalDataObject.get(entry.getKey());
-                        ServiceFunction originalServiceFunction = (ServiceFunction) dataObject;
-                        Object[] serviceFunctionObj = {originalServiceFunction};
-                        Class[] serviceFunctionClass = {ServiceFunction.class};
+                            ServiceFunction updatedServiceFunction = (ServiceFunction) entry.getValue();
 
+                            // We only update SF type entry if type has changed
+                            if (!updatedServiceFunction.getType().equals(originalServiceFunction.getType())) {
 
-                        ServiceFunction updatedServiceFunction = (ServiceFunction) entry.getValue();
-
-                        // We only update SF type entry if type has changed
-                        if (!updatedServiceFunction.getType().equals(originalServiceFunction.getType())) {
-
-                            // We remove the original SF from SF type list
-                            Future future = odlSfc.getExecutor().submit(SfcProviderServiceTypeAPI
-                                    .getDeleteServiceFunctionFromServiceType(serviceFunctionObj, serviceFunctionClass));
-                            try {
-                                LOG.debug("getDeleteServiceFunctionFromServiceType returns: {}", future.get());
-                            } catch (InterruptedException e) {
-                                LOG.warn("failed to ....", e);
-                            } catch (ExecutionException e) {
-                                LOG.warn("failed to ....", e);
+                                // We remove the original SF from SF type list
+                                Future future = odlSfc.getExecutor().submit(SfcProviderServiceTypeAPI
+                                        .getDeleteServiceFunctionFromServiceType(serviceFunctionObj, serviceFunctionClass));
+                                try {
+                                    LOG.debug("getDeleteServiceFunctionFromServiceType returns: {}", future.get());
+                                } catch (InterruptedException e) {
+                                    LOG.warn("failed to ....", e);
+                                } catch (ExecutionException e) {
+                                    LOG.warn("failed to ....", e);
+                                }
+                                // We create a independent entry
+                                serviceFunctionObj[0] = updatedServiceFunction;
+                                serviceFunctionClass[0] = ServiceFunction.class;
+                                odlSfc.getExecutor().submit(SfcProviderServiceTypeAPI
+                                        .getCreateServiceFunctionTypeEntry(serviceFunctionObj, serviceFunctionClass));
                             }
-                            // We create a independent entry
-                            serviceFunctionObj[0] = updatedServiceFunction;
-                            serviceFunctionClass[0] = ServiceFunction.class;
-                            odlSfc.getExecutor().submit(SfcProviderServiceTypeAPI
-                                    .getCreateServiceFunctionTypeEntry(serviceFunctionObj, serviceFunctionClass));
+
+                            /* Before removing RSPs used by this Service Function, we need to remove all
+                             * references in the SFF/SF operational trees
+                             */
+                            String sfName = originalServiceFunction.getName();
+                            List<SfServicePath> sfServicePathList = SfcProviderServiceFunctionAPI.
+                                    readServiceFunctionStateExecutor(sfName);
+                            List<String> rspList = new ArrayList<>();
+                            if ((sfServicePathList != null) && (!sfServicePathList.isEmpty())) {
+                                if (!SfcProviderServiceFunctionAPI.deleteServiceFunctionStateExecutor(sfName)) {
+                                    LOG.error("{}: Failed to delete SF {} operational state",
+                                            Thread.currentThread().getStackTrace()[1], sfName);
+                                }
+                                for (SfServicePath sfServicePath : sfServicePathList) {
+                                    String rspName = sfServicePath.getName();
+                                    SfcProviderServiceForwarderAPI
+                                            .deletePathFromServiceForwarderStateExecutor(rspName);
+                                    rspList.add(rspName);
+                                }
+                                SfcProviderRenderedPathAPI.deleteRenderedServicePathsExecutor(rspList);
+                            }
+                            /* We do not update the SFF dictionary. Since the user configured it in the first place,
+                             * (s)he is also responsible for updating it.
+                             */
                         }
-
-                /* Before removing RSPs used by this Service Function, we need to remove all
-                 * references in the SFF/SF operational trees
-                 */
-                        String sfName = originalServiceFunction.getName();
-                        List<SfServicePath> sfServicePathList = SfcProviderServiceFunctionAPI.
-                                readServiceFunctionStateExecutor(sfName);
-                        List<String> rspList = new ArrayList<>();
-                        if ((sfServicePathList != null) && (!sfServicePathList.isEmpty())) {
-                            if (!SfcProviderServiceFunctionAPI.deleteServiceFunctionStateExecutor(sfName)) {
-                                LOG.error("{}: Failed to delete SF {} operational state",
-                                        Thread.currentThread().getStackTrace()[1], sfName);
-                            }
-                            for (SfServicePath sfServicePath : sfServicePathList) {
-                                String rspName = sfServicePath.getName();
-                                SfcProviderServiceForwarderAPI
-                                        .deletePathFromServiceForwarderStateExecutor(rspName);
-                                rspList.add(rspName);
-                            }
-                            SfcProviderRenderedPathAPI.deleteRenderedServicePathsExecutor(rspList);
-                        }
-                /* We do not update the SFF dictionary. Since the user configured it in the first place,
-                 * (s)he is also responsible for updating it.
-                 */
                     }
                 }
             } finally {
